requirements.txt
holidays~=0.36
paho-mqtt~=1.6.1